requirements.txt
streamlit
faiss-cpu
openai
google-generativeai
tqdm
PyPDF2
python-dotenv
flask
flask-cors
sentence-transformers